rename: wish > wishAPI로 이름 변경
#117
